Add files via upload
In the third project I built and deployed a landmark image classification system using convolutional neural networks. Designed a CNN from scratch to process and classify a public landmark dataset, implemented the transfer learning with a pre-trained model (RNN-16) to improve performance. Exported the best-performing models using TorchScript to obtain a checkpoint that can be deployed in a user-facing app for real-time landmark identification. The project involved hands-on work in data preprocessing, model selection and tuning, and deployment.